{API breaking] Rename renderer and render cache #62
Renamed renderer to rasterizer and render cache to image atlas as part of #62.
The plan is to call the GPU rendering "renderer".

The rasterizer and image atlas were chose as they more closely describe the functionality.

In general anything "render_cache" is called "image_atlas" and anything "renderer" is "rasterizer".
The textures within the render cached used to be called images, now they are called textures.