Added culling for text layout rendering
Implements #14 #15

The culling bounds are meant to be used to allow to skip rendering for parts of the text. We already support logical bounds, but glyph geometry may extend beyond the logical boundries.

The culling could be done per quad, but I wanted to have a way to cull rendering before we even request the glyph, as it will also request the glyph to be rasterized.

Since calling skb_font_get_glyph_bounds() was not trivial, added cache for th eglyph bounds. Some approximation would be ok too, I experimented with calculating a common bounds for all the glyphs in a font, but some fonts (e.g. IBMPlexSansArabic) has wide glyphs which throws things off.

- Added culling bounds calculation for lines and runs
- Added common glyph bounds for runs, which is a bounding rect that can contain any of the glyphs of the run
- The common glyph bounds can be used for roughly culling glyphs, without having to store the glyph bounds per glyph (may revert this later if we end up needing glyph ink bounds for something else)
- Added render_draw_layout_with_culling which uses the above ink bounds to do line, run and glyph culling
- Some example code clean up