[API breaking] Change internal representation of the runs
This change is in preparation for rich text editing, consolidates some concepts and simplifies data access.

skb_attribute_span_t is removed, and the remaining data was moved to skb_layout_run_t. Now there are just different types of runs (content -> shaping -> layout) that flow through the layout build process.

The underlying representation now maintains runs along the whole layout process, which allowed to keep more data in the runs, making the glyph struct smaller, and simplified some algorithms.

The output of shaping is now array of shaping runs in logical order, array of glyphs in visual order, and array of clusters in logical order.

A cluster represents a range text that maps to range of glyphs that is inseparable.

The logical ordered shaping runs are turned into visual ordered layout runs when the text is laid out in lines. This change should also allow to reflow the lines without having to shape and analyze the text.

- removed skb_attribute_span_t and it's use
- moved following data from skb_attribute_span_t to skb_layout_run_t:
	- font_size
	- attribute slice
	- content_run_id
- store internal versions of the content runs
- added script, bidi_level, and index to content run to skb_layout_run_t
- make the shaping runs persistent (to later allow reflow of the lines without having to reshape)
- changed shaping to output clusters and glyphs (allows glyph struct to be slimmer)
- changed font selection to use cmap before shaping, and removed tricky fallback when a glyph is not found
- moved culling data to lines, since layout run struct is getting big
- removed glyph_range from skb_layout_line_t since it no longer is a linear range for the whole line (use layout runs instead)
- fixed skb__add_unique() which was not doing what it says on the tin

Migration:

Most use cases will encourter that skb_attribute_span_t is no more, so lines like this should be removed:
	<remove> const skb_attribute_span_t* attribute_span = &attribute_spans[run->attribute_span_idx]; </remove>

The font_size and attributes are now accessible via layout run, instead of attribute span:
	- attribute_span->font_size -> run->font_size
	- attribute_span->attributes -> run->attributes