[API breaking] Added paragraph style specific enter, tab and backspace handling
This change adds paragraph style specific handling for enter, tab and backspace.
For example, pressing enter at the end of a header row, will apply body style to the next row.

The goal is to move the special case handling to the editor, and maybe make it a bit customizable.
The code will live in the example, until few more rich text editing bits are completed.

- renamed skb_rich_layout_get_paragraph_position() -> skb_rich_layout_text_position_to_paragraph_position()
- renamed skb_editor_get_paragraph_text_offset() -> skb_editor_get_paragraph_global_text_offset()
- renamed skb_editor_get_text_offset_at() -> skb_editor_text_position_to_text_offset()
- added skb_editor_text_position_to_paragraph_position()
- added skb_editor_get_paragraph_text_count()
- fixed bug where skb_rich_text_set_paragraph_attribute_delta() would create new empty indent attribute

Migration:
- skb_rich_layout_get_paragraph_position() -> skb_rich_layout_text_position_to_paragraph_position()
- skb_editor_get_paragraph_text_offset() -> skb_editor_get_paragraph_global_text_offset()
- skb_editor_get_text_offset_at() -> skb_editor_text_position_to_text_offset()