Added group tag
Since we have only flat list of paragraphs, some formatting options cannot be expressed.

For example each item in a list is separate paragraph, but for good typography, we should be able to set the space before and after the list.

Group tag allows us to implicitly create the "parent container" based on the tag similarity. This allows some attributes to operate on the whole group.

Vertical alignment now has new member group_spacing, which is used to space items within the same group, and before/after paddings are used on the first and last item on the group.